Move snake_to_pascal back to root util
This is useful outside of EPICS contexts
