+++ src/fastcs/transports/epics/gui.py
@@ -36,7 +36,7 @@ from fastcs.logging import bind_logger
 from fastcs.methods import Command
 from fastcs.transports.controller_api import ControllerAPI
 from fastcs.transports.epics.options import EpicsGUIFormat, EpicsGUIOptions
-from fastcs.transports.epics.util import snake_to_pascal
+from fastcs.util import snake_to_pascal
 
 logger = bind_logger(logger_name=__name__)
