+++ src/fastcs/transports/epics/util.py
@@ -1,15 +1,5 @@
-import re
-
 from fastcs.transports.controller_api import ControllerAPI
-
-
-def snake_to_pascal(name: str) -> str:
-    """Converts string from snake case to Pascal case.
-    If string is not a valid snake case it will be returned unchanged
-    """
-    if re.fullmatch(r"[a-z][a-z0-9]*(?:_[a-z0-9]+)*", name):
-        name = re.sub(r"(?:^|_)([a-z0-9])", lambda match: match.group(1).upper(), name)
-    return name
+from fastcs.util import snake_to_pascal
 
 
 def controller_pv_prefix(prefix: str, controller_api: ControllerAPI) -> str:
